requirements.txt
requests
spidev
psycopg2
python-dotenv
azure-storage-blob
pyodbc
sqlalchemy
pydantic
azure-identity
